requirements.txt
dash==2.9.3
dash-bootstrap-components==1.7.1
dash-extensions==0.0.67
Django==5.0
django-appconf==1.1.0
django-compressor==4.5.1
django-crispy-forms==2.4
crispy-bootstrap4==2025.6
django-cors-headers==4.3.1
django-plotly-dash==2.3.1
django-tailwind==3.8.0
djangorestframework==3.15.2
dpd-static-support==0.0.5
joblib==1.3.2
keras==2.12.0
matplotlib==3.8.4
numpy==1.23.5
pandas==2.2.2
plotly==5.18.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2023.3.post1
scikit-learn==1.6.1
scipy==1.10.1
statsforecast==2.0.1
statsmodels==0.14.4
tailwind==3.1.5b0
tensorflow==2.12.0
tensorflow-cpu==2.12.0
whitenoise==6.6.0
gunicorn==21.2.0
channels==4.1.0
crispy_bootstrap4
